chore: Channel::join() 파일 분리
