+++ src/commands/join.cpp
@@ -5,7 +5,6 @@
 #include "ChannelService.hpp"
 
 typedef std::vector<std::string>::const_iterator str_iter;
-std::string RPL_JOIN(Session const &session, std::string const &channelName);
 bool isValidChannel(const std::string &channel);
 
 void ChannelService::join(Session *session, const Message &message) {
@@ -35,42 +34,9 @@ void ChannelService::join(Session *session, const Message &message) {
     }
 }
 
-void Channel::join(Session *session, const std::string &key) {
-    if (!isParticipant(session)) {
-        return NumericReply(ERR_USERONCHANNEL) << session << name >> session;
-    }
-
-    if (limit && participants.size() >= limit) {
-        return NumericReply(ERR_CHANNELISFULL) << session << name >> session;
-    }
-    if (inviteOnly) {
-        return NumericReply(ERR_INVITEONLYCHAN) << session << name >> session;
-    }
-    if (!channelKey.empty() && channelKey != key) {
-        return NumericReply(ERR_BADCHANNELKEY) << session << name >> session;
-    }
-
-    addParticipant(session);
-
-    *this << RPL_JOIN(*session, name);
-    if (!channelTopic.empty()) {
-        NumericReply(RPL_TOPIC, this->channelTopic) << session << name >> session;
-    }
-    NumericReply(RPL_NAMREPLY, getParticipantList()) << session << PUBLIC_CHANNEL_SYMBOL << name >> session;
-    NumericReply(RPL_ENDOFNAMES) << session << name >> session;
-}
-
 bool isValidChannel(const std::string &channel) {
     if (channel.length() < 2 || channel.size() > CHANNELLEN) {
         return false;
     }
     return (channel[0] == '#' || channel[0] == '&');
-}
-
-std::string RPL_JOIN(Session const &session, std::string const &channelName) {
-    std::stringstream ss;
-    ss << MESSAGE_PREFIX << session.getAddress() << DELIMITER
-       << "JOIN" << DELIMITER
-       << channelName << CRLF;
-    return ss.str();
 }
